conveyor_pi/pyproject.toml
[project]
name = "conveyor-pi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyserial",
    "flask>=3.0",
    "flask-socketio>=5.3",
    "numpy>=1.24",
    "Pillow>=10.0",
    "onnxruntime>=1.24.4",
]
